change wallet data type to bigint

+++ oas-server/oas/src/main/java/com/cascv/oas/server/blockchain/service/PowerService.java
@@ -1,16 +1,16 @@
 package com.cascv.oas.server.blockchain.service;
 
-import java.math.BigDecimal;
+import java.math.BigInteger;
 
 import org.springframework.stereotype.Service;
 
 @Service
 public class PowerService {
   public static final Integer powerFactor = 100;
-  public BigDecimal exchange(Integer power, Integer value) {
-    BigDecimal bigValue = BigDecimal.valueOf(value);
-    BigDecimal bigPower = BigDecimal.valueOf(power);
-    BigDecimal bigPowerFactor = BigDecimal.valueOf(powerFactor);
+  public BigInteger exchange(Integer power, Integer value) {
+    BigInteger bigValue = BigInteger.valueOf(value);
+    BigInteger bigPower = BigInteger.valueOf(power);
+    BigInteger bigPowerFactor = BigInteger.valueOf(powerFactor);
     return bigValue.multiply(bigPower).divide(bigPowerFactor);
   } 
 }
